+++ src/main/java/com/kakaoseventeen/dogwalking/notification/service/NotificationService.java
@@ -4,8 +4,8 @@ import com.kakaoseventeen.dogwalking.dog.Dog;
 import com.kakaoseventeen.dogwalking.dog.DogJpaRepository;
 import com.kakaoseventeen.dogwalking.member.domain.Member;
 import com.kakaoseventeen.dogwalking.notification.domain.Notification;
-import com.kakaoseventeen.dogwalking.notification.dto.LoadDogResponseDTO;
-import com.kakaoseventeen.dogwalking.notification.dto.LoadNotificationResponseDTO;
+import com.kakaoseventeen.dogwalking.notification.dto.response.LoadDogResponseDTO;
+import com.kakaoseventeen.dogwalking.notification.dto.response.LoadNotificationResponseDTO;
 import com.kakaoseventeen.dogwalking.notification.repository.NotificationJpaRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
